add deck db/router
