+++ src/server/auth.js
@@ -1,6 +1,6 @@
 const { GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET } = require('../../secrets.js');
 const passport = require('passport');
-const db = require('../db/db.js');
+const user = require('../db/userDb.js');
 
 // this function takes an OAuth profile and searches the database for
 // a matching user. If found, the function returns the user info. If
@@ -10,7 +10,7 @@ const db = require('../db/db.js');
 // google API on behalf of the user but at the moment I cannot thnk
 // of a use case
 const cb = async (request, accessToken, refreshToken, profile, done) => {
-  const userInfo = await db.getUser(profile.sub);
+  const userInfo = await user.getUser(profile.sub);
   // check if user is found
   if (userInfo) {
     // the first  argument of done is err. You must set err to null or else
@@ -20,7 +20,7 @@ const cb = async (request, accessToken, refreshToken, profile, done) => {
     // user not found so add user to db
     // profile._json contains the following fields:
     //   sub, picture, email, email_verified
-    const _id = await db.addUser(profile._json);
+    const _id = await user.addUser(profile._json);
     return done(null, { ...profile._json, _id });
   }
 };
